@@ -16,35 +16,40 @@ import org.springframework.transaction.annotation.Transactional;
 import com.andres.curso.springboot.app.springbootcrud.entities.User;
 import com.andres.curso.springboot.app.springbootcrud.repositories.UserRepository;
 
+
+// Este servicio es el que se encarga de explicarle a Spring security como funcionan los usuarios y roles en nuestra aplicacion
 @Service
 public class JpaUserDetailsService implements UserDetailsService {
 
     @Autowired
     private UserRepository repository;
 
-    @Transactional(readOnly = true)
+    @Transactional(readOnly = true) // Al decirle que es de tipo lectura optimizamos la consulta
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
 
         Optional<User> userOptional = repository.findByUsername(username);
 
+        // En caso de que no exista el usuario lanzamos una excepcion
         if (userOptional.isEmpty()) {
             throw new UsernameNotFoundException(String.format("Username %s no existe en el sistema!", username));
         }
 
         User user = userOptional.orElseThrow();
 
+        // Metemos en una lista 
         List<GrantedAuthority> authorities = user.getRoles().stream()
         .map(role -> new SimpleGrantedAuthority(role.getName()))
-                .collect(Collectors.toList());
+        .collect(Collectors.toList());
 
+        // Devolvemos el usuario
         return new org.springframework.security.core.userdetails.User(user.getUsername(), 
         user.getPassword(), 
         user.isEnabled(),
         true,
         true,
-        true,
-                authorities);
+                true,
+                        authorities);
     }
     
 }
