@@ -5,7 +5,7 @@ import javax.crypto.SecretKey;
 import io.jsonwebtoken.Jwts;
 
 public class TokenJwtConfig {
-    
+    // Esto son constantes, el token se genera de forma aleatoria al arrancar el proyecto
     public static final SecretKey SECRET_KEY = Jwts.SIG.HS256.key().build();
     public static final String PREFIX_TOKEN = "Bearer ";
     public static final String HEADER_AUTHORIZATION = "Authorization";
