@@ -8,9 +8,10 @@ import java.lang.annotation.Target;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 
+// Este Constraint significa que delega de la clase Original
 @Constraint(validatedBy = IsExistsDbValidation.class)
-@Target(ElementType.FIELD)
-@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.FIELD) // A que afecta
+@Retention(RetentionPolicy.RUNTIME) // significa que estará disponible en tiempo de ejecución
 public @interface IsExistsDb {
     String message() default "ya existe en la base de datos!";
 
